feat(core): extract conductor core package
