+++ packages/conductor-core/README.md
@@ -1,0 +1,9 @@
+# Conductor Core
+
+Reusable prompt-to-MIDI loop generation engine for Conductor.
+
+Install for local development:
+
+```bash
+pip install -e .
+```
