+++ packages/conductor-core/pyproject.toml
@@ -1,0 +1,45 @@
+[build-system]
+requires = ["setuptools>=69", "wheel"]
+build-backend = "setuptools.build_meta"
+
+[project]
+name = "conductor-core"
+version = "0.1.0"
+description = "Reusable prompt-to-MIDI loop generation engine for Conductor"
+readme = "README.md"
+requires-python = ">=3.10"
+dependencies = [
+	"mido==1.3.3",
+	"pydantic==2.11.4",
+]
+
+[project.optional-dependencies]
+openai = ["openai==1.79.0"]
+anthropic = ["anthropic==0.115.0"]
+google = ["google-genai==1.65.0"]
+ollama = ["ollama==0.6.1"]
+playback = [
+	"midi2audio==0.1.1",
+	"pydub==0.25.1",
+]
+providers = [
+	"anthropic==0.115.0",
+	"google-genai==1.65.0",
+	"ollama==0.6.1",
+	"openai==1.79.0",
+]
+dev = ["pytest>=8.3,<9"]
+
+[tool.setuptools]
+package-dir = {"" = "src"}
+include-package-data = true
+
+[tool.setuptools.packages.find]
+where = ["src"]
+
+[tool.setuptools.package-data]
+conductor_core = [
+	"resources/model_list.json",
+	"resources/prompts/loop_gen.txt",
+	"resources/soundfonts/*.sf2",
+]
