+++ packages/conductor-core/src/conductor_core/__init__.py
@@ -1,0 +1,19 @@
+"""Conductor Core public API."""
+
+from conductor_core.config import (
+    EngineConfig,
+    GenerationRequest,
+    GenerationResult,
+    ProgressEvent,
+    ProviderCredentials,
+)
+from conductor_core.engine import LoopGenerationEngine
+
+__all__ = [
+    "EngineConfig",
+    "GenerationRequest",
+    "GenerationResult",
+    "LoopGenerationEngine",
+    "ProgressEvent",
+    "ProviderCredentials",
+]
