+++ packages/conductor-core/src/conductor_core/resources/__init__.py
@@ -1,0 +1,1 @@
+"""Packaged default resources for Conductor Core."""
