+++ packages/conductor-core/src/conductor_core/resources/model_list.json
