+++ packages/conductor-core/src/conductor_core/resources/prompts/__init__.py
@@ -1,0 +1,1 @@
+"""Packaged prompt resources for Conductor Core."""
