+++ packages/conductor-core/src/conductor_core/resources/prompts/loop_gen.txt
