+++ packages/conductor-core/src/conductor_core/resources/soundfonts/__init__.py
@@ -1,0 +1,1 @@
+"""Packaged SoundFont resources for Conductor Core."""
