+++ src/claude_api.py
@@ -1,182 +1,7 @@
-from anthropic import Anthropic
-from dotenv import load_dotenv
-import src.objects as objects
-import src.utils as utils
-import logging
+"""Temporary compatibility alias for conductor_core.providers.anthropic."""
+
 import sys
-import os
 
-logging.basicConfig(level=logging.INFO, stream=sys.stderr, format='%(asctime)s - %(levelname)s - %(message)s')
-logger = logging.getLogger(__name__)
+from conductor_core.providers import anthropic as _anthropic
 
-ALWAYS_ON_ADAPTIVE_THINKING_MODELS = {"claude-fable-5", "claude-mythos-5"}
-ANTHROPIC_CACHE_CONTROL_MIN_CHARS = 4096
-
-def initialize_anthropic_client():
-    """
-    Initializes and returns an Anthropic client using API key from the .env file.
-    
-    Returns:
-        Anthropic: The Anthropic client instance.
-    """
-    load_dotenv(os.path.join("src", ".env"))
-    api_key = os.getenv('ANTHROPIC_API_KEY')
-    if not api_key or not api_key.strip():
-        logger.error("ANTHROPIC_API_KEY is not set!")
-    return Anthropic(api_key=api_key)
-
-def calc_price(model, output):
-    """
-    Calculate the cost for a given completion based on token usage.
-
-    Args:
-        model (str): The model used.
-        input_tokens (int): The input token count from the start event.
-        output_tokens (int): The total output token count.
-
-    Returns:
-        float: Calculated price for the API call.
-    """
-    model_info = utils.get_model_info()
-    if model not in model_info["models"]["Anthropic"].keys():
-        logger.warning(f"Model {model} not found in model info.")
-        return False
-    else:
-        input_cost = model_info["models"]["Anthropic"][model]["cost"]["input"] / 1000000
-        output_cost = model_info["models"]["Anthropic"][model]["cost"]["output"] / 1000000
-
-        cached_5min = model_info["models"]["Anthropic"][model]["cost"]["5m cache input"] / 1000000
-        cached_1hour = model_info["models"]["Anthropic"][model]["cost"]["1h cache input"] / 1000000
-        cache_hits = model_info["models"]["Anthropic"][model]["cost"]["cache hits/refreshes"] / 1000000
-                    
-        total_price = (output["input_tokens"] * input_cost) + (output["output_tokens"] * output_cost) + (output["cache_creation"] * cached_5min) + (output["cache_read"] * cache_hits)
-        return total_price
-
-
-def build_system_prompt_block(loop_prompt):
-    """Build Anthropic's system text block, marking only likely-cacheable prompts."""
-    block = {"type": "text", "text": loop_prompt}
-    if len(loop_prompt) >= ANTHROPIC_CACHE_CONTROL_MIN_CHARS:
-        block["cache_control"] = {"type": "ephemeral"}
-    return block
-
-def process_streaming_response(completion):
-    # Extract text, tool JSON, and token usage from the streaming response.
-    output = {
-        "loop": "",
-        "text": "",
-        "thinking_content": "",
-        "input_tokens": 0,
-        "output_tokens": 0,
-        "cache_creation": 0,
-        "cache_read": 0
-    }
-
-    # Process the streaming response
-    for chunk in completion:
-        if chunk.type == 'message_start':
-            if hasattr(chunk, "message") and hasattr(chunk.message, "usage"):
-                output["input_tokens"] += getattr(chunk.message.usage, "input_tokens", 0)
-                output["output_tokens"] += getattr(chunk.message.usage, "output_tokens", 0)
-                output["cache_creation"] += getattr(chunk.message.usage, "cache_creation_input_tokens", 0)
-                output["cache_read"] += getattr(chunk.message.usage, "cache_read_input_tokens", 0)
-        elif chunk.type == 'content_block_delta':
-            if hasattr(chunk.delta, "thinking"):
-                output["thinking_content"] += chunk.delta.thinking
-            elif hasattr(chunk.delta, "text"):
-                output["text"] += chunk.delta.text
-            elif hasattr(chunk.delta, "partial_json"):
-                output["loop"] += chunk.delta.partial_json
-        elif chunk.type == 'message_delta':
-            if hasattr(chunk, "usage"):
-                output["input_tokens"] += getattr(chunk.usage, "input_tokens", 0)
-                output["output_tokens"] += getattr(chunk.usage, "output_tokens", 0)
-                output["cache_creation"] += getattr(chunk.usage, "cache_creation_input_tokens", 0)
-                output["cache_read"] += getattr(chunk.usage, "cache_read_input_tokens", 0)
-        elif chunk.type == 'message_stop':
-            break
-    return output
-
-
-def loop_gen(prompt, model, temp=0.0, use_thinking=False, effort="low"):
-    """
-    Generate a MIDI bar (chord progression/melody) using the specified model and prompt.
-
-    Args:
-        prompt (str): The user prompt to generate MIDI data.
-        model (str): The model identifier to use.
-        temp (float, optional): Temperature for generation. Defaults to 0.0.
-        use_thinking (bool, optional): Enable extended thinking for supported models. Defaults to False.
-        effort (str, optional): Effort level for adaptive reasoning. Defaults to "low".
-
-    Returns:
-        tuple: (midi_loop, messages, cost)
-    """
-    # Initialize Anthropic client and make the API call
-    client = initialize_anthropic_client()
-    loop_prompt = utils.get_loop_prompt()
-    # Define the tool schema for the loop generation
-    loop_schema = objects.Loop.model_json_schema()
-    tools = [
-        {
-            "name": "build_MIDI_loop",
-            "description": "builds a music loop in MIDI format",
-            "input_schema": loop_schema
-        }
-    ]    
-    
-    # Prepare API call parameters
-    model_info = utils.get_model_info()
-    model_config = model_info["models"]["Anthropic"][model]
-    always_on_adaptive_thinking = model in ALWAYS_ON_ADAPTIVE_THINKING_MODELS
-    api_params = {
-        "model": model,
-        "max_tokens": model_config["max_tokens"],
-        "system": [build_system_prompt_block(loop_prompt)],
-        "messages": [{"role": "user", "content": prompt}],
-        "tools": tools,
-        "tool_choice": {"type": "tool", "name": "build_MIDI_loop"},
-        "stream": True
-    }
-    if not always_on_adaptive_thinking:
-        api_params["temperature"] = temp
-    
-    # Add thinking configuration
-    # Auto-enable adaptive thinking for models with effort_options when it is not always on.
-    # Thinking requires tool_choice "auto" (forced tool use not allowed)
-    if model_config.get("effort_options"):
-        api_params["tool_choice"] = {"type": "auto"}
-        if not always_on_adaptive_thinking:
-            api_params["thinking"] = {"type": "adaptive"}
-        api_params["output_config"] = {"effort": effort}
-        if not always_on_adaptive_thinking:
-            api_params["temperature"] = 1.0
-    # Legacy extended thinking requires explicit opt-in and auto tool_choice
-    elif use_thinking and model_config.get("extended_thinking"):
-        api_params["tool_choice"] = {"type": "auto"}
-        api_params["thinking"] = {
-            "type": "enabled",
-            "budget_tokens": model_config["max_thinking_budget"]
-        }
-        api_params["temperature"] = 1.0
-    elif use_thinking and not model_config.get("extended_thinking"):
-        logger.warning(f"Extended thinking requested but not supported by model: {model}")
-    
-    # Make the API call for structured output generation
-    completion = client.messages.create(**api_params)
-    output = process_streaming_response(completion)
-    if not output["loop"]:
-        raise ValueError(f"Model {model} did not call the build_MIDI_loop tool. Response text: {output['text'][:200]}")
-    loop = objects.Loop.model_validate_json(output["loop"])
-    
-    # Create the messages history 
-    messages = [
-        {"role": "system", "content": loop_prompt},
-        {"role": "user", "content": prompt}
-    ]
-    if output["thinking_content"]:
-        messages.append({"role": "assistant", "content": output["thinking_content"]})
-    messages.append({"role": "assistant", "content": output["loop"]})
-    
-    cost = calc_price(model, output)
-    return loop, messages, cost
+sys.modules[__name__] = _anthropic
